+++ src/misc.c
@@ -90,6 +90,33 @@ int ev_ipv6_name(const struct sockaddr_in6* addr, int* port, char* ip, size_t le
     return 0;
 }
 
+int ev_ip_addr(const char* ip, int port, struct sockaddr* addr, size_t size)
+{
+    if (strstr(ip, ":") != NULL)
+    {
+        if (size < sizeof(struct sockaddr_in6))
+        {
+            return EV_EINVAL;
+        }
+        return ev_ipv6_addr(ip, port, (struct sockaddr_in6*)addr);
+    }
+
+    if (size < sizeof(struct sockaddr_in))
+    {
+        return EV_EINVAL;
+    }
+    return ev_ipv4_addr(ip, port, (struct sockaddr_in*)addr);
+}
+
+int ev_ip_name(const struct sockaddr* addr, int* port, char* ip, size_t len)
+{
+	if (addr->sa_family == AF_INET)
+	{
+		return ev_ipv4_name((struct sockaddr_in*)addr, port, ip, len);
+	}
+	return ev_ipv6_name((struct sockaddr_in6*)addr, port, ip, len);
+}
+
 ev_buf_t ev_buf_make(void* buf, size_t len)
 {
     ev_buf_t tmp;